Added isolation jumpers, removed AC side PCB in favour of harness.
Added JP1/2 for isolating the output.  These can be renamed if these
labels are reserved for another jumper elsewhere in the project.

Thoughts here are that there could be a three way selector switch.
Off-Auto-On
The Off position would actuate a normally closed switch that opens in
this position.  The on switch would actuate a normally open switch that
closes to force the system on.  The auto selection in the centre
position actuates neither switch allowing the system to control itself.

Have added a back contact to simplify board design and preserve current
footprint.  Suggest moving JP1/2 to the right side of the board or
extending board width.

Also suggest a second set of Onewire contacts for a second external
connector so that modular units can be daisy chained.

Again board size may be extended to facilitate this but I am unsure
about footprint requirements.

Lastly, harness from AC side of SSR to board, then from board to socket
on enclosure seems pointless, simplify to a single harness with ring
lugs and just attach directly from input socket to SSR's and output
sockets.